requirements.txt
authlib
Flask
flask_sqlalchemy
pydantic
PyJWT
python-dotenv
Werkzeug #Я НЕ ЕБУ ЧТО ЖЭТО
psycopg2-binary
dotenv
flasgger